[삼성 SDS] 수신자 구분 필터 조회, 장비 트래픽 통계 이슈 수정
- 메세지 조회 > 수신자 세부화 (상세) > 수신자 구분 필터 조회 결과값 안나오는 문제 수정
-> 엘라스틱 필드와 UI 검색 필드 불일치 문제로 엘라스틱 필드와 동일하게 맞춤

- 장비 트래픽 통계 조회 불가 문제 수정
-> 트래픽 사이즈를 INT 형으로 정의해두었으나, INT 형의 최대값이 2,147,483,647로 제한 , 고객사의 트래픽 사이즈가 이 값을 초과할 경우 0으로 출력
-> 트래픽 사이즈를 INT 형이 아닌 LONG 형으로 변경

+++ src/main/java/com/xcurenet/device/service/DeviceTrafficStatVO.java
@@ -8,7 +8,7 @@ public class DeviceTrafficStatVO {
 	private String updateDt;
 	private String port;
 	private int direction;
-	private int hourSize;
+	private long hourSize;
 	private int hourCnt;
 	private String insertDt;
 }
